#12039 Refactor
