+++ backend/de.metas.adempiere.adempiere/base/src/main/java/de/metas/location/CountryService.java
@@ -22,6 +22,7 @@
 
 package de.metas.location;
 
+import de.metas.location.CountryDisplaySequenceHelper;
 import lombok.NonNull;
 import org.compiere.model.I_C_Country;
 import org.compiere.model.I_C_Country_Sequence;
